examples/hello-world/offchain/requirements.txt
pycardano==0.9.0
python-dotenv==1.0.0
requests==2.31.0
